+++ src/node/ext/ldap/tests.py
@@ -55,4 +55,4 @@ def test_suite():
     return suite
 
 if __name__ == '__main__':
-    unittest.main(defaultTest='test_suite')                # pragma NO COVERAGE
+    unittest.main(defaultTest='test_suite')                #pragma NO COVERAGE