+++ src/node/ext/ldap/ugm/_api.py
@@ -114,11 +114,8 @@ class PrincipalAliasedAttributes(object):
 
     def __init__(self, context, aliaser=None):
         """
-        context
-            the node whose children to alias
-
-        aliaser
-            the aliaser to be used
+        :param context: The node whose children to alias
+        :param aliaser: The aliaser to be used
         """
         self.__name__ = context.name
         self.__parent__ = None
@@ -586,9 +583,9 @@ class LDAPPrincipals(OdictStorage):
         return unaliased_dct
 
     @default
-    def search(self, criteria=None, attrlist=None,
-               exact_match=False, or_search=False, or_keys=None,
-               or_values=None, page_size=None, cookie=None):
+    def raw_search(self, criteria=None, attrlist=None,
+                   exact_match=False, or_search=False, or_keys=None,
+                   or_values=None, page_size=None, cookie=None):
         search_attrlist = [self._key_attr]
         if attrlist is not None and self._key_attr not in attrlist:
             search_attrlist += attrlist
@@ -623,6 +620,25 @@ class LDAPPrincipals(OdictStorage):
         if cookie is not None:
             return results, cookie
         return results
+
+    @default
+    def search(self, criteria=None, attrlist=None,
+               exact_match=False, or_search=False):
+        result = []
+        cookie = None
+        while True:
+            chunk, cookie = self.raw_search(
+                criteria=criteria,
+                attrlist=attrlist,
+                exact_match=exact_match,
+                or_search=or_search,
+                page_size=self.context.ldap_session._props.page_size,
+                cookie=cookie
+            )
+            result += chunk
+            if not cookie:
+                break
+        return result
 
     @default
     @locktree
@@ -992,23 +1008,12 @@ class LDAPUgm(UgmBase):
     def __init__(self, name=None, parent=None, props=None,
                  ucfg=None, gcfg=None, rcfg=None):
         """
-        name
-            node name
-
-        parent
-            node parent
-
-        props
-            LDAPProps
-
-        ucfg
-            UsersConfig
-
-        gcfg
-            GroupsConfig
-
-        rcfg
-            RolesConfig
+        :param name: Node name.
+        :param parent: Node parent.
+        :param props: LDAPProps instance.
+        :param ucfg: UsersConfig instance.
+        :param gcfg: GroupsConfig instance.
+        :param rcfg: RolesConfig instance.
         """
         self.__name__ = name
         self.__parent__ = parent